# apps/api-client-go/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@boxlite.com
    name: BoxLite Platforms Inc.
    url: https://www.boxlite.io
  description: BoxLite AI platform API Docs
  title: BoxLite
  version: "1.0"
servers:
- url: http://localhost:3000
paths:
...
  /preview/{boxId}/public:
    get:
      operationId: isBoxPublic
      parameters:
      - description: ID of the box
        explode: false
        in: path
        name: boxId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                type: boolean
          description: Public status of the box
      summary: Check if box is public
      tags:
      - preview
  /preview/{boxId}/validate/{authToken}:
    get:
      operationId: isValidAuthToken
      parameters:
      - description: ID of the box
        explode: false
        in: path
        name: boxId
        required: true
        schema:
          type: string
        style: simple
      - description: Auth token of the box
        explode: false
        in: path
        name: authToken
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                type: boolean
          description: Box auth token validation status
      summary: Check if box auth token is valid
      tags:
      - preview
  /preview/{boxId}/access:
    get:
      operationId: hasBoxAccess
      parameters:
      - explode: false
        in: path
        name: boxId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                type: boolean
          description: User access status to the box
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Check if user has access to the box
      tags:
      - preview
...
components:
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      description: API Key access
      scheme: bearer
      type: http
    oauth2:
      openIdConnectUrl: http://localhost:3000/.well-known/openid-configuration
      type: openIdConnect